requirements.txt
Pillow==8.3.2
python-telegram-bot==13.5
pandas==1.3.3


